:anchor: Manage listings page, user authorization

+++ app/Http/Controllers/ListingController.php
@@ -64,6 +64,12 @@ class ListingController extends Controller
 
     public function update(Request $request, Listing $listing) {
 
+        // Make sure logged in user is owner
+
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -90,8 +96,20 @@ class ListingController extends Controller
 
     public function destroy(Listing $listing) {
 
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
+
+    }
+
+    // Manage Listings
+
+    public function manage() {
+
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
 
     }
 }
